requirements.txt
python-telegram-bot[job-queue]==21.6
python-dotenv==1.0.1
openai==1.52.2
dateparser==1.2.0
httpx==0.27.2
